implement search and modified as per feedback
