+++ src/book.js
@@ -2,7 +2,7 @@ const fs = require('fs'); //like import
 
 //1. Read data from json file (contains array type)
 let rawdata = fs.readFileSync('./src/data/books.json');  
-let books = JSON.parse(rawdata);  //returns js arrays
+let books = JSON.parse(rawdata);  //returns js array
 
 //returns list of books 
 function getBookList(){
@@ -34,10 +34,54 @@ function getBookById(bookId){
     }
     return book;
 }
+
+
+//search books by author ID 
+function searchBookByAuthorID(AuthorID){
+    var booksByAuthor = [];
+    for(var i = 0; i < books.length; i++){
+        if (AuthorID == books[i].author.id){
+            booksByAuthor.push(books[i]);
+        }
+
+    }
+    if (booksByAuthor.length == 0){
+        return "The author ID provided is not in the file.";
+    }
+    else{
+        return booksByAuthor;
+    }
+    
+}
  
-// module.exports is an object that the current module returns when it is "required" in another program or module.
-// it is like package 
 module.exports = {
-    getBookList: getBookList, // you can assign any variable as key 
-    getBookById: getBookById
+    searchBookByAuthorID: searchBookByAuthorID
+};
+
+
+//search books by keyword 
+function searchBookByKeyWord(keyWord){
+    var listOfBooks = [];
+    for(var i = 0; i < books.length; i++){
+        if (books[i].title.search(keyWord) != -1 || books[i].description.search(keyWord) != -1){
+            listOfBooks.push(books[i]);
+        }
+
+    }
+    if (listOfBooks.length == 0){
+        return "There is no books associated with this keyword.";
+    }
+    else{
+        return listOfBooks;
+    }
+    
+}
+ 
+ 
+ 
+module.exports = {
+    getBookList: getBookList,  
+    getBookById: getBookById,
+    searchBookByAuthorID: searchBookByAuthorID,
+    searchBookByKeyWord: searchBookByKeyWord
 };
